M3-ize common/ui/compose
