--- common/ui/compose/src/main/java/app/tivi/common/compose/theme/AppBarAlphas.kt
@@ -1,29 +1,0 @@
-/*
- * Copyright 2021 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package app.tivi.common.compose.theme
-
-import androidx.compose.material.MaterialTheme
-import androidx.compose.runtime.Composable
-
-object AppBarAlphas {
-    @Composable
-    fun translucentBarAlpha(): Float = when {
-        // We use a more opaque alpha in light theme
-        MaterialTheme.colors.isLight -> 0.97f
-        else -> 0.94f
-    }
-}
