+++ common/ui/compose/src/main/java/app/tivi/common/compose/theme/Color.kt
@@ -16,19 +16,9 @@
 
 package app.tivi.common.compose.theme
 
-import androidx.compose.material.Colors
-import androidx.compose.material.darkColors
-import androidx.compose.material.lightColors
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.compositeOver
-
-/**
- * Return the fully opaque color that results from compositing [onSurface] atop [surface] with the
- * given [alpha]. Useful for situations where semi-transparent colors are undesirable.
- */
-fun Colors.compositedOnSurface(alpha: Float): Color {
-    return onSurface.copy(alpha = alpha).compositeOver(surface)
-}
 
 internal val Slate200 = Color(0xFF81A9B3)
 internal val Slate600 = Color(0xFF4A6572)
@@ -37,22 +27,16 @@ internal val Slate800 = Color(0xFF232F34)
 internal val Orange500 = Color(0xFFF9AA33)
 internal val Orange700 = Color(0xFFC78522)
 
-val TiviLightColors = lightColors(
+val TiviLightColors = lightColorScheme(
     primary = Slate800,
     onPrimary = Color.White,
-    primaryVariant = Slate600,
     secondary = Orange700,
-    secondaryVariant = Orange500,
     onSecondary = Color.Black
 )
 
-val TiviDarkColors = darkColors(
+val TiviDarkColors = darkColorScheme(
     primary = Slate200,
     onPrimary = Color.Black,
     secondary = Orange500,
     onSecondary = Color.Black
-).withBrandedSurface()
-
-fun Colors.withBrandedSurface() = copy(
-    surface = primary.copy(alpha = 0.08f).compositeOver(this.surface)
 )
