--- common/ui/compose/src/main/java/app/tivi/common/compose/theme3/Theme.kt
@@ -1,47 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package app.tivi.common.compose.theme3
-
-import android.os.Build
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.dynamicDarkColorScheme
-import androidx.compose.material3.dynamicLightColorScheme
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalContext
-
-@Composable
-fun TiviTheme(
-    useDarkColors: Boolean = isSystemInDarkTheme(),
-    content: @Composable () -> Unit
-) {
-    MaterialTheme(
-        colorScheme = when {
-            Build.VERSION.SDK_INT >= 31 && useDarkColors -> {
-                dynamicDarkColorScheme(LocalContext.current)
-            }
-            Build.VERSION.SDK_INT >= 31 && !useDarkColors -> {
-                dynamicLightColorScheme(LocalContext.current)
-            }
-            useDarkColors -> TiviDarkColors
-            else -> TiviLightColors
-        },
-        typography = TiviTypography,
-        shapes = TiviShapes,
-        content = content
-    )
-}
