+++ common/ui/compose/src/main/java/app/tivi/common/compose/ui/PosterCard.kt
@@ -18,14 +18,12 @@ package app.tivi.common.compose.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Card
-import androidx.compose.material.ContentAlpha
-import androidx.compose.material.LocalContentAlpha
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
+import androidx.compose.material3.Card
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
@@ -46,17 +44,13 @@ fun PosterCard(
         Box(
             modifier = if (onClick != null) Modifier.clickable(onClick = onClick) else Modifier
         ) {
-            // TODO: remove text if the image has loaded (and animated in).
-            // https://github.com/chrisbanes/accompanist/issues/76
-            CompositionLocalProvider(LocalContentAlpha provides ContentAlpha.medium) {
-                Text(
-                    text = show.title ?: "No title",
-                    style = MaterialTheme.typography.caption,
-                    modifier = Modifier
-                        .padding(4.dp)
-                        .align(Alignment.CenterStart)
-                )
-            }
+            Text(
+                text = show.title ?: "No title",
+                style = MaterialTheme.typography.bodySmall,
+                modifier = Modifier
+                    .padding(4.dp)
+                    .align(Alignment.CenterStart)
+            )
             if (poster != null) {
                 AsyncImage(
                     model = poster,
@@ -65,7 +59,7 @@ fun PosterCard(
                         UiR.string.cd_show_poster_image,
                         show.title ?: "show"
                     ),
-                    modifier = Modifier.matchParentSize(),
+                    modifier = Modifier.fillMaxSize(),
                     contentScale = ContentScale.Crop
                 )
             }
