+++ common/ui/compose/src/main/java/app/tivi/common/compose/ui/RefreshButton.kt
@@ -18,10 +18,10 @@ package app.tivi.common.compose.ui
 
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.size
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Refresh
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
