+++ common/ui/compose/src/main/java/app/tivi/common/compose/ui/TiviAlertDialog.kt
@@ -16,10 +16,10 @@
 
 package app.tivi.common.compose.ui
 
-import androidx.compose.material.AlertDialog
-import androidx.compose.material.OutlinedButton
-import androidx.compose.material.Text
-import androidx.compose.material.TextButton
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 
 @Composable
