+++ ui/account/src/main/java/app/tivi/account/AccountUi.kt
@@ -17,11 +17,11 @@
 package app.tivi.account
 
 import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -33,6 +33,7 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.Divider
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Surface
@@ -44,13 +45,11 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import app.tivi.common.compose.theme.foregroundColor
 import app.tivi.common.compose.ui.AsyncImage
 import app.tivi.data.entities.TraktUser
 import app.tivi.trakt.TraktAuthState
@@ -102,7 +101,8 @@ internal fun AccountUi(
 ) {
     Surface(
         color = MaterialTheme.colorScheme.surfaceVariant,
-        shape = MaterialTheme.shapes.medium
+        shape = MaterialTheme.shapes.medium,
+        modifier = Modifier.defaultMinSize(minHeight = 200.dp)
     ) {
         Column {
             Spacer(modifier = Modifier.height(16.dp))
@@ -222,10 +222,9 @@ private fun AppAction(
     ) {
         Spacer(modifier = Modifier.width(8.dp))
 
-        Image(
+        Icon(
             imageVector = icon,
-            contentDescription = contentDescription,
-            colorFilter = ColorFilter.tint(foregroundColor())
+            contentDescription = contentDescription
         )
 
         Spacer(modifier = Modifier.width(16.dp))
